requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0 